Added more i18n titles on the cv
Also removed the paragraph of About me because that's actually
data. It's static on this site but it'd be dynamic on a real
context so it's better and more organized to keep the assumption
that's dynamic (even when it's not). It simplifies things and
prepares the code for making it dynamic if needed.