Patched @react-pdf/pdfkit 3.0.2 bug
So seems that some fonts cause issues when they are copied and
pasted. And ATS software reads text that way probably so I needed
to fix it. Otherwise I'd be stuck with default fonts.

I've created this patch package based on this pending PR
https://github.com/diegomura/react-pdf/pull/2408 that fixes
the issue so now I can use any font!