Fixed different height animation problem
Depending on the lines of text the animation would push the
container higher or lower, resulting in cut content.
By animating max-height we can always get the animation correct
because the actual height is dynamic. However we need to make
sure the hover max-height will always be enough, and also we
cannot make it too tall or the animation will reproduce very
slow.