Created ProjectGalleryPressable component
This is so I can abstract the logic of building the necessary
props for the LightGalleryPressable.

I do this with a new component because I think it's preferable
to keep the LightGalleryPressable component pure so I can reuse
it easily for other gallery types rather than binding it to the
Project type permanently.