Made compatible with or without base
Now I can add a base if I want or not, it's flexible.
Before without a base was crashing. This is useful
if I want to server from the root of a domain but then
for some reason if I want to serve it from a folder I just
can add it.