Making English page the root one
This is to help SEO basically, otherwise the root URL only
has a JS script without any tags and it's not good.

This introduces a bug that if a user wants to visit
the English he cannot unless he has the browser set
in English. Which is not desirable, every user should
be able to see whatever page he wants. I'll think of
a fix.

This invalidates the /en/ URL but I'll create a
redirect version later.