Executing the redirection script before HTML is downloaded
This may seem unimportant, but now I have reduced by a factor of three
the speed on which the redirection happens. On slow connections
this is important because maybe the HTML was being downloaded, but
since the user had another language, it had to be redirected.
And then the HTML would download again.
So it's very very simple and yields very very useful results.