Bolding a word in about me section
Also, I'm disabling the inline script tag because, although it was
fixing a rendering issue, it was impossible to use more than one instance
of the element. And now that we need more than one we need this disabled.
For now it's better this way and we will find a fix later on.