Hiding arrows on gallery depending on the slide
This is a good thing because sometimes the image gets obscured by the
arrow and we minimize this occurrence if we hide one arrow.

Also, for whatever reason the LightGallery property that should work
isn't working so I have to code it myself.

Docs: https://www.lightgalleryjs.com/docs/settings/#hideControlOnEnd